backend/Authentication_and_Users/auth-service/app/utils/jwt_handler.py
import jwt
from datetime import datetime, timedelta
import os

SECRET_KEY = os.getenv("JWT_SECRET", "secreto123")

def create_jwt_token(data: dict, expires_delta=timedelta(hours=1)):
    to_encode = data.copy()
    to_encode["exp"] = datetime.utcnow() + expires_delta
    return jwt.encode(to_encode, SECRET_KEY, algorithm="HS256")

