backend/Authentication_and_Users/auth-service/requirements.txt
fastapi
uvicorn
pyjwt
python-jose
pydantic
psycopg2-binary